requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.0
pydantic-settings>=2.5.2

# Multi-Agent Framework
crewai>=0.86.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
litellm>=1.80.0
google-generativeai>=0.8.0

# GitHub Integration
PyGithub>=2.4.0
httpx>=0.27.2

# Database
motor>=3.6.0
pymongo>=4.9.0,<4.10
beanie>=1.26.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
authlib>=1.3.2
itsdangerous>=2.2.0

# Utilities
python-dotenv>=1.0.1
pytz>=2024.2
aiofiles>=24.1.0

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx==0.27.2
